dgds/pyproject.toml
[tool.poetry]
name = "dgds"
version = "0.1.0"
description = ""
authors = ["BLooperZ <blooperz@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pakal = {git = "https://github.com/adventurebrew/pakal.git"}
numpy = "^2.0.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
